requirements.txt
flask
requests
sentence-transformers
faiss-cpu
langchain-community
hf_xet
